Refactor: Move email service to email_service module
- Rename finbot/core/email/ to finbot/core/email_service/
- Update auth module to import from new location
- Improves code organization and naming consistency
