--- finbot/core/email/__init__.py
@@ -1,13 +1,0 @@
-"""Email service module for magic link authentication"""
-
-from finbot.core.email.base import EmailService
-from finbot.core.email.console import ConsoleEmailService
-from finbot.core.email.factory import get_email_service
-from finbot.core.email.resend_client import ResendEmailService
-
-__all__ = [
-    "EmailService",
-    "ConsoleEmailService",
-    "ResendEmailService",
-    "get_email_service",
-]
