+++ finbot/core/email_service/base.py
